+++ drift/test/platforms/ffi/native_database_test.dart
@@ -1,13 +1,40 @@
 @TestOn('vm')
+import 'dart:io';
+
 import 'package:drift/drift.dart';
 import 'package:drift/native.dart';
 import 'package:sqlite3/sqlite3.dart';
 import 'package:test/test.dart';
+import 'package:test_descriptor/test_descriptor.dart' as d;
 
+import '../../generated/todos.dart';
 import '../../test_utils/database_vm.dart';
 
 void main() {
   preferLocalSqlite3();
+
+  group('implicit isolates', () {
+    test('work with direct executors', () async {
+      final file = File(d.path('test.db'));
+
+      final db = TodoDb(NativeDatabase.createInBackground(file));
+      await db.todosTable.select().get(); // Open the database
+      await db.close();
+
+      await d.file('test.db', anything).validate();
+    });
+
+    test('work with connections', () async {
+      final file = File(d.path('test.db'));
+
+      final db =
+          TodoDb.connect(NativeDatabase.createBackgroundConnection(file));
+      await db.todosTable.select().get(); // Open the database
+      await db.close();
+
+      await d.file('test.db', anything).validate();
+    });
+  });
 
   group('NativeDatabase.opened', () {
     test('disposes the underlying database by default', () async {